Link player  with cztenis nonregistered user

+++ src/main/java/cz/cvut/fel/tk21/service/InvitationService.java
@@ -1,6 +1,7 @@
 package cz.cvut.fel.tk21.service;
 
 import cz.cvut.fel.tk21.dao.InvitationDao;
+import cz.cvut.fel.tk21.exception.ValidationException;
 import cz.cvut.fel.tk21.model.Club;
 import cz.cvut.fel.tk21.model.Invitation;
 import cz.cvut.fel.tk21.model.User;
@@ -8,6 +9,7 @@ import cz.cvut.fel.tk21.model.UserRole;
 import cz.cvut.fel.tk21.model.mail.Mail;
 import cz.cvut.fel.tk21.service.mail.MailService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,13 +21,15 @@ public class InvitationService extends BaseService<InvitationDao, Invitation>{
     private final MailService mailService;
     private final ClubRelationService clubRelationService;
     private final UserService userService;
+    private final PasswordEncoder passwordEncoder;
 
     @Autowired
-    protected InvitationService(InvitationDao dao, MailService mailService, ClubRelationService clubRelationService, UserService userService) {
+    protected InvitationService(InvitationDao dao, MailService mailService, ClubRelationService clubRelationService, UserService userService, PasswordEncoder passwordEncoder) {
         super(dao);
         this.mailService = mailService;
         this.clubRelationService = clubRelationService;
         this.userService = userService;
+        this.passwordEncoder = passwordEncoder;
     }
 
     public Optional<Invitation> findByConfirmationToken(String token){
@@ -43,6 +47,8 @@ public class InvitationService extends BaseService<InvitationDao, Invitation>{
             token = UUID.randomUUID().toString();
         }
         invite.setConfirmationToken(token);
+
+        this.persist(invite);
         return invite;
     }
 
@@ -51,18 +57,39 @@ public class InvitationService extends BaseService<InvitationDao, Invitation>{
         User user = invitation.getUser();
         Club club = invitation.getClub();
 
+        if(!user.isVerifiedAccount()) throw new ValidationException("Uživatel není ověřen");
+
         if(clubRelationService.isMemberOf(club, user)){
             clubRelationService.deleteRoleWithoutPermissionCheck(club, user, UserRole.RECREATIONAL_PLAYER);
             clubRelationService.addRoleWithoutPermissionCheck(club, user, UserRole.PROFESSIONAL_PLAYER);
             //TODO uncomment
             //sendInvitationAcceptedMail(invitation);
         } else {
-            //TODO implement
+            clubRelationService.addUserToClub(club, user, UserRole.PROFESSIONAL_PLAYER);
         }
 
         user.setWebId(invitation.getWebId());
         userService.update(user);
         this.remove(invitation);
+    }
+
+    @Transactional
+    public void changePasswordAndAcceptInvitation(Invitation invitation, String password){
+        User user = invitation.getUser();
+        Club club = invitation.getClub();
+
+        if(user.isVerifiedAccount()) throw new ValidationException("Uživatel je již ověřen");
+
+        user.setPassword(passwordEncoder.encode(password));
+        user.setVerifiedAccount(true);
+        user.setWebId(invitation.getWebId());
+        userService.update(user);
+        this.remove(invitation);
+
+        clubRelationService.addUserToClub(club, user, UserRole.PROFESSIONAL_PLAYER);
+
+        //TODO uncomment
+        //sendInvitationAcceptedMail(invitation);
     }
 
     public void sendProfessionalPlayerInviteMail(Invitation invitation){
@@ -76,6 +103,19 @@ public class InvitationService extends BaseService<InvitationDao, Invitation>{
         mail.setModel(model);
 
         mailService.sendProfessionalPlayerInvite(mail);
+    }
+
+    public void sendProfessionalPlayerInviteMailNonRegisteredPlayer(Invitation invitation){
+        Mail mail = new Mail();
+        mail.setFrom("noreply@tk21.cz");
+        mail.setTo(invitation.getUser().getEmail());
+        mail.setSubject("Pozvání do aplikace TK21");
+
+        Map<String, Object> model = new HashMap<>();
+        model.put("token", invitation.getConfirmationToken());
+        mail.setModel(model);
+
+        mailService.sendProfessionalPlayerInviteNonRegisteredPlayer(mail);
     }
 
     public void sendInvitationAcceptedMail(Invitation invitation){
